projetos/process-ia-mvp/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.2.5
google-generativeai==0.8.3
email-validator==2.2.0
alembic==1.13.3
